fix: time type

+++ internal/heartbeat/new.go
@@ -40,11 +40,11 @@ type Heartbeat struct {
 	LinesInFile    *int    `json:"lines,omitempty"`
 	Project        *string `json:"project,omitempty"`
 	ProjectPath    *string `json:"projectPath,omitempty"`
-	Time           float64 `json:"time"`
+	Time           int64   `json:"time"`
 	UserAgent      string  `json:"userAgent"`
 }
 
-func New(entity, userAgent string, time float64, cursorPos *int, lang *string, lineNum *int, linesInFile *int, project *string, projectPath *string) *Heartbeat {
+func New(entity, userAgent string, time int64, cursorPos *int, lang *string, lineNum *int, linesInFile *int, project *string, projectPath *string) *Heartbeat {
 	hb := &Heartbeat{
 		Entity:         entity,
 		Time:           time,
